+++ nodedump.js
@@ -19,15 +19,17 @@ var hljs = require('./lib/highlight.js/highlight');
 
 // Default options
 var DEFAULTOPTS = {
-	expand: true
-	,label:null
-	,show:null
+	collapse: false
+	,dumpFunctionName: 'nodedump'
+	,expand: true
 	,hide:null
-	,top:null
+	,hideTypes:null
+	,label:null
 	,levels: null
+	,show:null
 	,sortKeys: true
 	,syntaxHighlight:true
-	,dumpFunctionName: 'nodedump'
+	,top:null
 };
 
 // used to figure out the datatype of a variable
@@ -132,7 +134,7 @@ var JS = "<script type=\"text/javascript\">\n\
 			} // end toggleRow\n\
 \n\
 			,toggleSource: function(source){\n\
-				if (source.style.fontStyle=='italic') {\n\
+				if (source.style.fontStyle == 'italic') {\n\
 					source.style.fontStyle='normal';\n\
 					source.title='" + TITLEEXPANDED + "';\n\
 					return 'open';\n\
@@ -164,7 +166,7 @@ var JS = "<script type=\"text/javascript\">\n\
 			} // end toggleTable\n\
 \n\
 			,toggleTarget: function(target,switchToState){\n\
-				target.style.display = (switchToState=='open') ? '' : 'none';\n\
+				target.style.display = (switchToState == 'open') ? '' : 'none';\n\
 			} // end toggleTarget\n\
 		};\n\
 \n\
@@ -187,67 +189,45 @@ function doTable(dataType, data){
 }
 
 /*
- * Checks if the current row is expanded or not
- * 
- * @param {object} options
- * @param {boolean} isSimpleTypeMember
- * @returns {isRowExpanded.options|Boolean}
- */
-function isRowExpanded(options, isSimpleTypeMember){
-	return (!isSimpleTypeMember || (options && options.expand == true));
-}
-
-/*
  * Builds the style tag for the headers of tables
  * 
  * @param {string} dataType
- * @param {object} options
+ * @param {Boolean} expand
  * @returns {String|EXPANDEDLABELSTYLE|COLLAPSEDLABELSTYLE}
  */
-function doHeaderStyle(dataType, options){
-	var style = EXPANDEDLABELSTYLE;
-	if(options.expand==false || (typeof options.expand=='object' && options.expand.indexOf(dataType) < 0))
-		style = COLLAPSEDLABELSTYLE;
-
-	return style;
+function doHeaderStyle(dataType, expand){
+	return expand ? EXPANDEDLABELSTYLE : COLLAPSEDLABELSTYLE;
 }
 
 /*
  * Builds the style tag for a row
  * 
  * @param {string} dataType
- * @param {object} options
- * @param {boolean} isSimpleTypeMember
+ * @param {Boolean} expand
  * @returns {COLLAPSEDSTYLE|String}
  */
-function doRowStyle(dataType, options, isSimpleTypeMember){
-	return ((
-			isRowExpanded(options, isSimpleTypeMember)
-			|| (typeof options.expand=='object' && options.expand.indexOf(dataType) > -1)
-			) ? '' : COLLAPSEDSTYLE
-	);
+function doRowStyle(dataType, expand){
+	return expand ? '' : COLLAPSEDSTYLE;
 }
 
 /*
  * Builds the style tag for the label cell
  * 
- * @param {string} options
- * @param {boolean} isSimpleTypeMember
+ * @param {Boolean} expand
  * @returns {String|COLLAPSEDLABELSTYLE|EXPANDEDLABELSTYLE}
  */
-function doCellLabelStyle(options, isSimpleTypeMember){
-	return isRowExpanded(options, isSimpleTypeMember) ? EXPANDEDLABELSTYLE : COLLAPSEDLABELSTYLE;
+function doCellLabelStyle(expand){
+	return expand ? EXPANDEDLABELSTYLE : COLLAPSEDLABELSTYLE;
 }
 
 /*
  * Builds the style tag for the data cell
  * 
- * @param {object} options
- * @param {boolean} isSimpleTypeMember
+ * @param {Boolean} expand
  * @returns {String|COLLAPSEDSTYLE}
  */
-function doCellStyle(options, isSimpleTypeMember){
-	return isRowExpanded(options, isSimpleTypeMember) ? '' : COLLAPSEDSTYLE;
+function doCellStyle(expand){
+	return expand ? '' : COLLAPSEDSTYLE;
 }
 
 /*
@@ -255,11 +235,11 @@ function doCellStyle(options, isSimpleTypeMember){
  * 
  * @param {string} dataType
  * @param {string} data
- * @param {object} options
+ * @param {Boolean} expand
  * @returns {string}
  */
-function doRowHeader(dataType, data, options){
-	return format(ROWHEADER, dataType, doHeaderStyle(dataType, options), data);
+function doRowHeader(dataType, data, expand){
+	return format(ROWHEADER, dataType, doHeaderStyle(dataType, expand), data);
 }
 
 /*
@@ -268,18 +248,18 @@ function doRowHeader(dataType, data, options){
  * @param {string} dataType
  * @param {string} key
  * @param {string} data
- * @param {object} options
- * @param {boolean} isSimpleTypeMember
+ * @param {Boolean} expand
+ * @param {Boolean} expandCell
  * @returns {string}
  */
-function doRow(dataType, key, data, options, isSimpleTypeMember){
+function doRow(dataType, key, data, expand, expandCell){
     return format(
 		ROW
-		, doRowStyle(dataType, options, isSimpleTypeMember)
+		, doRowStyle(dataType, expand)
 		, dataType
-		, doCellLabelStyle(options, isSimpleTypeMember)
+		, doCellLabelStyle(expandCell)
 		, key
-		, doCellStyle(options, isSimpleTypeMember)
+		, doCellStyle(expandCell)
 		, data
 	);
 }
@@ -289,23 +269,22 @@ function doRow(dataType, key, data, options, isSimpleTypeMember){
  * 
  * @param {string} dataType
  * @param {string} data
- * @param {object} options
+ * @param {Boolean} expand
  * @returns {string}
  */
-function doRowHeader1Col(dataType, data, options){
-    return format(ROWHEADER1COL, dataType, doHeaderStyle(dataType, options), data);
+function doRowHeader1Col(dataType, data, expand){
+    return format(ROWHEADER1COL, dataType, doHeaderStyle(dataType, expand), data);
 }
 
 /*
  * Builds the 1 column row
  * @param {string} dataType
  * @param {string} data
- * @param {object} options
- * @param {boolean} isSimpleTypeMember
+ * @param {Boolean} expand
  * @returns {string}
  */
-function doRow1Col(dataType, data, options, isSimpleTypeMember){
-    return format(ROW1COL, doRowStyle(dataType, options, isSimpleTypeMember), data);
+function doRow1Col(dataType, data, expand){
+    return format(ROW1COL, doRowStyle(dataType, expand), data);
 }
 
 /*
@@ -438,6 +417,9 @@ function dumpObject(obj, cache, currentPath, options){
 	var isSimpleType = (SIMPLETYPES.indexOf(dataType) >= 0);
 	var bFirstCall = (currentPath.length == 0);
 	var label = dataType;
+	var expand = true;
+	var expandCells = true;
+
 	if(bFirstCall){
 		var topPath = TOP;
 		cache.bFilteredLevel = false;
@@ -459,7 +441,7 @@ function dumpObject(obj, cache, currentPath, options){
 		switch(dataType){
 			case 'Boolean':
 				var val = '<span class="'+(obj ? 'nodedump-yes' : 'nodedump-no')+'">' + obj + '</span>';
-				data = doRow(dataType, label, val, options);
+				data = doRow(dataType, label, val, expand, expandCells);
 			break;
 			case 'String':
 				if(obj.length === 0)
@@ -467,30 +449,44 @@ function dumpObject(obj, cache, currentPath, options){
 				else {
 					var val = escapeHtml(obj);
 					//var val = '<pre><code class="lang-html">' + hljs.highlight('xml', obj).value + '</code></pre>';
-					data = doRow(dataType, label, val, options);
+					data = doRow(dataType, label, val, expand, expandCells);
 				}
 			break;
 			case 'Math':
 			case 'Undefined':
 			case 'Null':
-				data = doRow1Col(dataType, label, options, false);
+				data = doRow1Col(dataType, label, expand);
 			break;
 			default:
-				data = doRow(dataType, label, obj.toString(), options);
+				data = doRow(dataType, label, obj.toString(), expand, expandCells);
 			break;
 		}
 
-	} else { // Non-Object types
+	} else { // Non-Simple types
+
+		// figure out if it should be expanded/collapsed
+		expand = options.expand;
+		if(typeof options.expand == 'object'){
+			expand = expand.indexOf(dataType) > -1 || expand.indexOf(dataType.toLowerCase()) > -1;
+		}
+		if(expand){
+			if(options.collapse == true || (
+					typeof options.collapse == 'object' 
+					&& (options.collapse.indexOf(dataType) > -1 || options.collapse.indexOf(dataType.toLowerCase()) > -1)
+				)
+			)
+				expand = false;
+		}
 
 		switch(dataType){
 			case 'RegExp':
 			case 'Error':
-				data += doRowHeader1Col(dataType, label, options);
-				data += doRow1Col(dataType, obj.toString(), options, true);
+				data += doRowHeader1Col(dataType, label, expand);
+				data += doRow1Col(dataType, obj.toString(), expand);
 			break;
 			case 'Function':
 				bHeader = true;
-				data += doRowHeader1Col(dataType, label, options);
+				data += doRowHeader1Col(dataType, label, expand);
 				var txt = obj.toString();
 				if(options.syntaxHighlight){
 					var purtyText = hljs.highlight('javascript', txt);
@@ -500,23 +496,28 @@ function dumpObject(obj, cache, currentPath, options){
 					var txt = escapeHtml(obj.toString());
 				}
 
-				data += doRow1Col(dataType, '<pre><code class="lang-javascript">'+txt+'</code></pre>', options, true);
-				//data += doRow1Col(dataType, '<pre><code>'+escapeHtml(obj.toString())+'</code></pre>', options, true);
+				data += doRow1Col(dataType, '<pre><code class="lang-javascript">'+txt+'</code></pre>', expand);
+				//data += doRow1Col(dataType, '<pre><code>'+escapeHtml(obj.toString())+'</code></pre>', expand);
 			break;
 			case 'Array':
 			case 'Object':
 			default:
+				// set keys to collapse if an array of types was passed for expand and the current data-type is one of them
+				expandCells = expand;
+				if(typeof options.expand == 'object' && expand)
+					expandCells = false;
+
 				// check for circular references
 				var circPath = getPathToCircularRef(obj, cache, currentPath);
 				if(circPath.length > 0){
 					dataType = CIRCULARREFERENCE;
-					data = doRow(dataType, dataType, circPath.join(CIRCULARSPLITSTRING), options);
+					data = doRow(dataType, dataType, circPath.join(CIRCULARSPLITSTRING), expand);
 				} else {
 					var subPath;
 					var loopObj = [];
 					for(var key in obj)
 						loopObj.push(key);
-					if(options.sortKeys){
+					if(dataType != 'Array' && options.sortKeys){
 						loopObj.sort(function (a, b) {
 							return a.toLowerCase().localeCompare(b.toLowerCase());
 						});
@@ -539,10 +540,10 @@ function dumpObject(obj, cache, currentPath, options){
 							errThrown = err.toString();
 						}
 						if(bFirstCall){
-							if(!(!options.show || (options.show.length && options.show.indexOf(key) >= 0))){
+							if(!(!options.show || (options.show.length && (options.show.indexOf(key) >= 0 || options.show.indexOf(Number(key)) >= 0)))){
 								numKeysHidden++;
 								continue;
-							} else if(options.hide && options.hide.length && options.hide.indexOf(key) >= 0){
+							} else if(options.hide && options.hide.length && (options.hide.indexOf(key) >= 0 || options.hide.indexOf(Number(key)) >= 0)){
 								numKeysHidden++;
 								continue;
 							}
@@ -551,24 +552,33 @@ function dumpObject(obj, cache, currentPath, options){
 								break;
 							}
 						}
+						// skip any data types that should be hidden
+						if(options.hideTypes){
+							var subDataType = getDataType(val);
+							if(options.hideTypes.indexOf(subDataType) > -1 || options.hideTypes.indexOf(subDataType.toLowerCase()) > -1){
+								numKeysHidden++;
+								continue;
+							}
+						}
 
 						numKeysShown++;
 						if(options.levels !== null && currentPath.length > options.levels){
 							cache.bFilteredLevel = true;
-							data += doRow(dataType, key, '', options, true);
+							data += doRow(dataType, key, '', true);
 							continue;
 						}
 						
 						if(errThrown.length > 0){
-							var errorRow = doRowHeader1Col(ERRORDATATYPE, ERRORDATATYPE, options)
-											+ doRow1Col(ERRORDATATYPE, '<pre><code class="lang-javascript">'+hljs.highlight('javascript', errThrown).value+'</code></pre>', options, true);
-											//+ doRow1Col(ERRORDATATYPE, errThrown, options, true);
-							data += doRow(dataType, key, doTable(ERRORDATATYPE, errorRow), options, false);
+							var errorRow = doRowHeader1Col(ERRORDATATYPE, ERRORDATATYPE, true)
+											+ doRow1Col(ERRORDATATYPE, '<pre><code class="lang-javascript">'+hljs.highlight('javascript', errThrown).value+'</code></pre>', true);
+											//+ doRow1Col(ERRORDATATYPE, errThrown, true);
+							data += doRow(dataType, key, doTable(ERRORDATATYPE, errorRow), expandCells);
 							continue;
 						}
 						subPath = clone(currentPath, 'Array');
 						subPath.push(key);
-						data += doRow(dataType, key, dumpObject(val, cache, subPath, options), options, true);
+
+						data += doRow(dataType, key, dumpObject(val, cache, subPath, options), expand, expandCells);
 					}
 					
 					if(numTotalKeys === 0)
@@ -576,22 +586,26 @@ function dumpObject(obj, cache, currentPath, options){
 					else {
 						if(bFirstCall){
 							if(numKeysShown !== numTotalKeys){
-								if(options.show){
+								if(options.show || options.hideTypes){
 									filtered.push(format(TITLEFILTEREDSHOWN, numKeysShown, numTotalKeys));
 								} else if(options.hide){
 									filtered.push(format(TITLEFILTEREDHIDDEN, numKeysHidden, numTotalKeys));
 									numTotalKeys = numTotalKeys - numKeysHidden;
 								}
-								if(!options.show && bFilteredTop)
+								if(!(options.show || options.hideTypes) && bFilteredTop)
 									filtered.push(format(TITLEFILTEREDTOP, numKeysShown, numTotalKeys));
 							}
 							if(cache.bFilteredLevel)
 								filtered.push(format(TITLEFILTEREDLEVELS, options.levels));
 
-							if(filtered.length > 0)
-								label += format(TITLEFILTERED, filtered.join(', '));
+						} else if(options.hideTypes && numKeysShown !== numTotalKeys){
+							// show the filtered label for types being hidden
+							filtered.push(format(TITLEFILTEREDSHOWN, numKeysShown, numTotalKeys));
 						}
-						data = doRowHeader(dataType, label, options) + data;
+						if(filtered.length > 0)
+							label += format(TITLEFILTERED, filtered.join(', '));
+
+						data = doRowHeader(dataType, label, expand) + data;
 					}
 				}
 			break;