+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+# 0.0.8
+- Added the ability to skip over data types like Functions when dumping: https://github.com/ragamufin/nodedump/pull/8
+- Fixed issue array keys not being sorted properly: https://github.com/ragamufin/nodedump/pull/8#issuecomment-47699217
+- Added new options 'hideTypes' and 'collapse': https://github.com/ragamufin/nodedump#options
+
 # 0.0.7
 - Fixed issue with path to highlight.js: https://github.com/ragamufin/nodedump/issues/7
+
 # 0.0.4
 - added this document
 - fix for arrays with non-numeric keys issue: https://github.com/ragamufin/nodedump/issues/5